+++ CHANGELOG.md
@@ -15,6 +15,8 @@ Please mark backwards incompatible changes with an exclamation mark at the start
   when `active_support/core_ext/string` hadn't been loaded.
 
 ### Added
+- The `Elasticsearch::Indices::Settings::Blocks` class. The class encapsulates
+  an index's blocks settings (for example, whether the index is read-only).
 - The `Elasticsearch::Indices::Settings` class. The class encapsulates an
   index's settings.
 - It is now possible to configure the type used by the `RSpec::TestDataCollector`